BPE26-ref: extract store subscriptions to dedicated file
